pyproject.toml
[project]
name = "avanza-skill"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["avanza-api"]

[dependency-groups]
test = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
